Add files via upload
Got minimization working with constraints! (almost). I still need to implement the parallel operator constraints, but it works if I force all the operators to be series, which is pretty cool. The issue with parallel is its a variable number of operands, so I have to find a workaround since I cant iterate over the operator value. Seems like its running fast though