Add files via upload
It works! In general, it can solve the problem for ~6 or fewer resistors in a matter of seconds. Still has some bugs. It doesnt find a solution for resistor sets of size 1 or 2. It still has the issue of the wrong implementation to check for duplicate resistor use if there is more than one of the same value. Theres a good amount to be done with optimization/guessing too.