Add files via upload
added some optimization through symmetry breaking so its faster now. Should still add more if I can think of any. Added more tests that all pass. Now if there is a tie for accuracy it defaults to the shorter solution. It also seems like it doesnt work for targets over 1000 but I have not looked that closely at this